requirements.txt
pytest
dotenv
google-genai